+++ snuba_sdk/metrics_visitors.py
@@ -1,7 +1,8 @@
 from __future__ import annotations
 
 from abc import ABC, abstractmethod
-from typing import Any, Generic, Mapping, TypeVar
+from dataclasses import asdict
+from typing import Any, Generic, Mapping, Sequence, TypeVar, Union
 
 from snuba_sdk.column import Column
 from snuba_sdk.conditions import (
@@ -284,6 +285,16 @@ class RollupSnQLPrinter(RollupVisitor[Mapping[str, str]]):
         }
 
 
+class RollupMQLPrinter(RollupVisitor[Mapping[str, Union[str, int, None]]]):
+    def visit(self, rollup: Rollup) -> dict[str, str | int | None]:
+        return {
+            "orderby": rollup.orderby.value if rollup.orderby else None,
+            "granularity": rollup.granularity,
+            "interval": rollup.interval,
+            "with_totals": str(rollup.totals) if rollup.totals is not None else None,
+        }
+
+
 class ScopeVisitor(ABC, Generic[TVisited]):
     @abstractmethod
     def visit(self, scope: MetricsScope) -> TVisited:
@@ -316,6 +327,11 @@ class ScopeSnQLPrinter(ScopeVisitor[str]):
         )
 
         return self.translator.visit(condition)
+
+
+class ScopeMQLPrinter(ScopeVisitor[Mapping[str, Union[str, Sequence[int]]]]):
+    def visit(self, scope: MetricsScope) -> dict[str, str | list[int]]:
+        return asdict(scope)
 
 
 # Normally this would be a visitor class, but this is itself a hack to get derived metrics off the ground