+++ snuba_sdk/mql_context.py
@@ -1,12 +1,10 @@
 from __future__ import annotations
 
-from dataclasses import dataclass, fields
-from datetime import datetime
-from typing import Any, Mapping, Sequence
+from dataclasses import dataclass
 
-from snuba_sdk.expressions import Limit, Offset
-from snuba_sdk.mql_context_visitors import MQLContextPrinter, Validator
-from snuba_sdk.timeseries import MetricsScope, Rollup
+
+class InvalidMQLContextError(Exception):
+    pass
 
 
 @dataclass
@@ -15,35 +13,30 @@ class MQLContext:
     The MQL string alone is not enough to fully describe a query.
     This class contains all of the additional information needed to
     execute a metrics query in snuba.
+
+    It should be noted that this class is used as an intermediary encoding
+    class for data in the the MetricsQuery class that can't be encoded into
+    MQL. As such it shouldn't be used directly by users of the SDK.
+
+    This also means that the validation here is quite loose, since this object
+    should be created exclusively from a valid MetricsQuery object.
     """
 
-    entity: str | None = None
-    start: datetime | None = None
-    end: datetime | None = None
-    rollup: Rollup | None = None
-    scope: MetricsScope | None = None
-    limit: Limit | None = None
-    offset: Offset | None = None
-    indexer_mappings: dict[str, Any] | None = None
+    entity: str
+    start: str
+    end: str
+    rollup: dict[str, str | int | None]
+    scope: dict[str, str | list[int]]
+    indexer_mappings: dict[str, str | int]
+    limit: int | None = None
+    offset: int | None = None
 
-    def get_fields(self) -> Sequence[str]:
-        self_fields = fields(self)
-        return tuple(f.name for f in self_fields)
+    def __post_init__(self) -> None:
+        self.validate()
 
     def validate(self) -> None:
-        # For now, we cannot validate entity because it unknown when
-        # we converting MetricsQuery to MQL. In that specific case, we
-        # need to set the entity on the requesst after serialization.
-
-        # In the future, we should be able to remove entity from this class
-        # entirely when we join entities together.
-        VALIDATOR.visit(self)
-
-    def serialize(self) -> Mapping[str, Any]:
-        self.validate()
-        result = MQL_CONTEXT_PRINTER.visit(self)
-        return result
-
-
-MQL_CONTEXT_PRINTER = MQLContextPrinter()
-VALIDATOR = Validator()
+        # Simple assert that all the expected fields are present
+        fields = ["entity", "start", "end", "rollup", "scope", "indexer_mappings"]
+        for field in fields:
+            if getattr(self, field) is None:
+                raise InvalidMQLContextError(f"MQLContext.{field} is required")